pc - refactor code to update roster students with user information by matching emails
This was in the GithubSignInServiceImpl, but it doesn't really belong there.
Now it's in its own service, and there's an sync job to run it on all users.
Eventually, that job should not be needed, once we figure out where in the architecture
this should go to ensure that the RosterStudent <-> User links are established at
the correct moments in the object lifecycle.