pc - rewrite scripts and instructions for converting the private key from the Github app
In this commit, we rewrite the scripts and the instructions
for installing the Github App.   Instead of passing the name of
the private key file, we just find it automatically; if there is
more than one, we prompt the user to select one.

We have separate scripts for localhost and dokku.

The localhost version just creates the secrets.yaml file
directly (warning the user if it is overwriting another one).

The dokku version gives the exact command to copy/paste
to the dokku command line.

The instructions have been modified to match these changes t
to the scripts, and have been streamlined to avoid
duplicating instructions that already appear elsewhere.